requirements.txt
fastapi
uvicorn
requests
python-dotenv
aws-requests-auth
